feat: move directory location
