+++ apps/web/src/common/modules/navigations/top-bar/modules/gnb-search-clone/type.ts
